S5-UC3 : Implemented the MySQL database and performed CRUD operations on it

+++ src/main/java/com/bridgelabz/AddressBookApp/exception/AddressNotFound.java
@@ -2,7 +2,7 @@ package com.bridgelabz.AddressBookApp.exception;
 
 public class AddressNotFound extends Exception {
 
-    private static final long serialVersionUID = 1L;
+    private static final int serialVersionUID = 1;
 
     public AddressNotFound(String message) {
         super(message);
